clean up session management
