+++ app/lib/session.server.ts
@@ -1,6 +1,7 @@
-import { createCookieSessionStorage } from '@remix-run/node';
+import { createCookieSessionStorage, redirect } from '@remix-run/node';
 
-// TODO: make this legit
+import { combineHeaders } from '~/lib/headers.server';
+
 export const sessionStorage = createCookieSessionStorage({
   cookie: {
     httpOnly: true,
@@ -11,3 +12,49 @@ export const sessionStorage = createCookieSessionStorage({
     secure: process.env.NODE_ENV === 'production'
   }
 });
+
+export async function requireAuth(request: Request) {
+  const session = await sessionStorage.getSession(
+    request.headers.get('Cookie')
+  );
+  const user = session.get('user');
+  if (!user) {
+    throw redirect('/');
+  }
+  return { user };
+}
+
+export async function login(
+  request: Request,
+  credentials: { email: string; password: string }
+) {
+  // TODO: validate credentials
+  await new Promise((resolve) => setTimeout(resolve, 1000));
+
+  const session = await sessionStorage.getSession(
+    request.headers.get('Cookie')
+  );
+  session.set('user', {
+    email: credentials.email
+  });
+
+  throw redirect('/dashboard', {
+    headers: {
+      'Set-Cookie': await sessionStorage.commitSession(session)
+    }
+  });
+}
+
+export async function logout(request: Request, responseInit?: ResponseInit) {
+  const session = await sessionStorage.getSession(
+    request.headers.get('Cookie')
+  );
+
+  throw redirect('/', {
+    ...responseInit,
+    headers: combineHeaders(
+      { 'Set-Cookie': await sessionStorage.destroySession(session) },
+      responseInit?.headers
+    )
+  });
+}
